header: fix signature when protected header is zero-length map
COSE allows an empty protected header to be encoded as a zero-length
map, even though the standard encorages encoding empty protected headers
as a zero-length string (RFC 2119 SHOULD according to RFC 8152,
Section 3).

However, according to RFC 8152, Section 4.4, even if the header is
encoded as a zero-length map, the signature data should not include the
empty map if the protected header is empty.

This commit ensures that this behavior is implemented in coset, which
previously did include the zero length map (encoded as h'a0') in
signature calculation.
This previously caused signature verification failures, e.g. when
verifying the CoseSign1 object provided in
https://github.com/cose-wg/Examples/blob/master/sign1-tests/sign-pass-03.json
using coset.